pyproject.toml
[tool.poetry]
name = "linkedin-collector"
version = "0.1.0"
description = ""
authors = ["creotip <creotip@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
playwright = "^1.41.2"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
asyncer = "^0.0.5"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-playwright = "^0.4.4"
pytest-asyncio = "^0.23.5"
ruff = "^0.3.0"
faker = "^23.3.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.deployer.dependencies]
boto3 = "^1.34.46"
generic-service-chart = {git = "https://github.com/revrod/generic-service-chart", rev = "main"}
typer = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    "N", # pep8-naming
    # isort
    "I001",
    "I002",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
